feature/validate

+++ database/migrations/2022_11_05_182426_create_sum_vinyl_table.php
@@ -22,8 +22,8 @@ return new class extends Migration
                 ->onDelete('cascade')
                 ->onUpdate('cascade');
 
-            $table->integer('width'); //Inches
-            $table->integer('height'); //Inches
+            $table->decimal('width', 8, 2); //Inches
+            $table->decimal('height', 8, 2); //Inches
             $table->decimal('price', 8, 2);
         });
     }
